requirements.txt
# Security Key - Dependências Python
# FastAPI e servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ORM e banco de dados
sqlalchemy>=2.0.0
alembic>=1.12.0

# Validação e configuração
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Autenticação
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Criptografia
cryptography>=41.0.0

# HTTP (para o OAuth2)
python-multipart>=0.0.6

# Segurança extra (opcional)
# argon2-cffi>=23.1.0

# Desenvolvimento (opcional)
# pytest>=7.4.0
# httpx>=0.25.0
